requirements.txt
bittensor==6.9.3
starlette>=0.30.0
pydantic>=2
rich>=13
pytest>=8
torch>=2
numpy>=1
setuptools>=68
langchain
langchain-openai
python-dotenv